Title: 두 정수에 대한 연산값, Time: 129ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 250225/두 정수에 대한 연산값/operational-values-for-two-integers.java
@@ -10,8 +10,8 @@ public class Main {
     }
 
     public static void swap(int a, int b) {
-        int s = 2* (int)Math.min(a,b);
-        int x = (int)Math.max(a,b)+25;
+        int s = 2 * Math.min(a,b);
+        int x = Math.max(a,b)+25;
         System.out.print(s+" "+x);
     }
 }
